requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
apscheduler==3.10.4
python-dotenv==1.0.0
httpx==0.24.1
streamlit==1.26.1
requests==2.31.0
